feat: add cluster/metrics endpoint (#717)
1. add: new endpoint /cluster/metrics
2. add: prometheus metrics parser
3. chore: code cleanup
