+++ server/Cargo.toml
@@ -104,6 +104,7 @@ serde_repr = "0.1.17"
 hashlru = { version = "0.11.0", features = ["serde"] }
 path-clean = "1.0.1"
 prost = "0.12.3"
+prometheus-parse = "0.2.5"
 
 [build-dependencies]
 cargo_toml = "0.15"
